% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcop.R
\name{dcop}
\alias{dcop}
\title{Density of a Gaussian Copula Distribution}
\usage{
dcop(x, par, dens.x, dens.y, Rho, mean, sd, log = FALSE, decomp = FALSE,
  max.Z = 10, debug = FALSE)
}
\arguments{
\item{x}{An array}

\item{par}{paramaters generated by \code{cop.par}}

\item{dens.x}{used as a parameter to \code{cop.par} if par if missing}

\item{dens.y}{used as a parameter to \code{cop.par} if par if missing}

\item{Rho}{used as a parameter to \code{cop.par} if par if missing}

\item{mean}{used as a parameter to \code{cop.par} if par if missing}

\item{sd}{used as a parameter to \code{cop.par} if par if missing}

\item{log}{A boolean (\code{FALSE} by default). If \code{TRUE}, this function will return the log of the density.}

\item{decomp}{A boolean (\code{FALSE} by default). If \code{TRUE}, this function will return a list containing
the marginally Normalized data z, its multivariate log-density ldens.z, and the log-jacobian ljac.z,
such that the total log-density is ldens.z + ljac.z
instead of returning the density as a single array}

\item{debug}{a boolean (\code{FALSE} by default) if set to \code{TRUE}, will cause the function to open a browser mid-call}

\item{max.z}{An integer. Normals will be truncated to +/- \code{max.z} standard deviations (default 10)}
}
\value{
the density as a numeric array
}
\description{
Calculates the density of a gaussian copula distribution.
}

