% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgcir.R
\name{dgcir}
\alias{dgcir}
\title{gcir stationary distribution (regular scale)}
\usage{
dgcir(x, gamma, mu, sigma, lambda, log = FALSE, debug = FALSE)
}
\arguments{
\item{debug}{a boolean (\code{FALSE} by default) if set to \code{TRUE}, will cause the function to open a browser mid-call}
}
\value{
a numeric vector
}

