% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhest.R
\name{dhest}
\alias{dhest}
\title{Stationary Distribution of Heston Model Log-Return Differences}
\usage{
dhest(x, alpha, gamma, mu, beta, sigma, rho, dT, inter = c(-5000, 5000),
  n = 1000, LConst, debug = FALSE)
}
\value{
numeric
}
\description{
these will roughly be centered at alpha*dT \cr
            requires Fourier inversion of the characteristic function, ie. numerical \cr
            complex integration.  currently done by grid method. \cr
            LConst allows fewer exp evaluations if alpha is held constant. \cr
}

