% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdeLists.R
\docType{data}
\name{ghestList}
\alias{ghestList}
\title{Ghest Model Parameters}
\format{a list containing ndims, nparams, param.names, data.names, sdeDr, sdeDf, isValidData, isValidParams}
\usage{
ghestList
}
\description{
List of paramaters that can be passed to sde.make.model to produce a ghest model.
}
\keyword{datasets}

