% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdeLists.R
\docType{data}
\name{gl05List}
\alias{gl05List}
\title{gl05 Model Parameters}
\format{a list containing ndims, nparams, param.names, data.names, sdeDr, sdeDf, isValidData, isValidParams}
\usage{
gl05List
}
\description{
List of paramaters that can be passed to sde.make.model to produce a gl05 model.
}
\keyword{datasets}

