% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msde-package.R
\docType{data}
\name{hestList}
\alias{hestList}
\title{Heston Model Parameters}
\format{A list containing ndims, nparams, param.names, data.names, sdeDr, sdeDf, isValidData, isValidParams}
\description{
List of paramaters that can be passed to sde.make.model to produce a ghest model
}
\keyword{datasets}

