% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{Methods}
\alias{Methods}
\alias{Robj.eval}
\alias{Methods}
\alias{Robj.set.x}
\alias{Methods}
\alias{Robj.get.x}
\alias{Methods}
\alias{sde.size}
\title{R Methods}
\usage{
Robj.eval(obj, y)

Robj.set.x(obj, x)

Robj.get.x(obj)

sde.size(model)
}
\description{
R Methods
}
