% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdeHeaders.R
\docType{package}
\name{msdeHeaders}
\alias{msdeHeaders}
\alias{msdeHeaders-package}
\title{Simulation and inference for multivariate stochastic differential equations.}
\description{
Simulation and inference for multivariate stochastic differential equations.
}
