% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msde-package.R
\name{rcop}
\alias{rcop}
\title{Simulation of a Gaussian Copula Distribution}
\arguments{
\item{n}{An integer giving the number of rows to be generated}

\item{par}{paramaters generated by \code{cop.par}}

\item{dens.x}{used as a parameter to \code{cop.par} if par if missing}

\item{dens.y}{used as a parameter to \code{cop.par} if par if missing}

\item{Rho}{used as a parameter to \code{cop.par} if par if missing}

\item{mean}{used as a parameter to \code{cop.par} if par if missing}

\item{sd}{used as a parameter to \code{cop.par} if par if missing}

\item{debug}{a boolean (\code{FALSE} by default) if set to \code{TRUE}, will cause the function to open a browser mid-call}
}
\value{
a matrix of size \code{n} x \code{length(dens.x)}
}
\description{
NOTE: problematic treatment of duplicate values in Dens.y.
            Only seems to be a problem for highly skewed marginals, so far...
}

