% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.diff.R
\name{sde.diff}
\alias{sde.diff}
\title{SDE Diffusion Function}
\usage{
sde.diff(model, x, theta, debug = FALSE)
}
\arguments{
\item{model}{An \code{sde.model} object.}

\item{x}{A vector or matrix of data with \code{ndims} columns.}

\item{theta}{A vector or matrix of parameters with \code{nparams} columns.}
}
\value{
A matrix with \code{ndims^2} columns containing the diffusion funtion evaluated at \code{x} and \code{theta}.  Each row corresponds to the upper triangular cholesky factor of the diffusion matrix.  If either input contains invalid sde data or parameters an error is thrown.
}
\description{
SDE Diffusion Function
}
