% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.drift.R
\name{sde.drift}
\alias{sde.drift}
\title{SDE Drift Function}
\usage{
sde.drift(model, x, theta, debug = FALSE)
}
\arguments{
\item{model}{An \code{sde.model} object.}

\item{x}{A vector or matrix of data with \code{ndims} columns.}

\item{theta}{A vector or matrix of parameters with \code{nparams} columns.}
}
\value{
A matrix with \code{ndims} columns containing the drift funtion evaluated at \code{x} and \code{theta}.  If either input contains invalid sde data or parameters an error is thrown.
}
\description{
SDE Drift Function
}
