% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.examples.R
\name{sde.examples}
\alias{sde.examples}
\title{Example SDE models.}
\usage{
sde.examples(model = c("hest", "pgnet", "lotvol", "biou"),
  ModelFile.only = FALSE)
}
\arguments{
\item{model}{Character string giving the name of a sample model.  Possible values are: \code{hest}, \code{pgnet}, \code{lotvol}, \code{biou}.  See Details.}

\item{ModelFile.only}{If \code{TRUE} returns only the path to the header file containing the \code{sdeModel} object implementation.}
}
\value{
An \code{sde.model} object (see \code{\link{sde.make.model}}).
}
\description{
Provides sample \code{C++} code for several SDE models.
}
\details{
A full description of the sample models can be found in the package vignette; to view it run \code{vignette("msde-exmodels")}.
}
\examples{
\donttest{
# Heston's model
hmod <- sde.examples("hest") # load pre-compiled model

# inspect model's C++ code
hfile <- example.models("hest")
cat(readLines(hfile), sep = "\\n")

# compile it from scratch
params.names <- c("alpha", "gamma", "beta", "sigma", "rho")
data.names <- c("X", "Z")
hmod <- sde.make.model(ModelFile = hfile,
                       param.names = param.names,
                       data.names = data.names)
}
}
