% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.loglik.R
\name{sde.loglik}
\alias{sde.loglik}
\title{SDE Loglikelihood Function}
\usage{
sde.loglik(model, x, dt, theta, ncores = 1, debug = FALSE)
}
\arguments{
\item{model}{An \code{sde.model} object.}

\item{x}{A matrix or 3-d array of data with \code{dim(x)[1]} observations and \code{dim(x)[2] == ndims}.}

\item{dt}{A scalar or vector of length \code{dim(x)[1]-1} of time intervals between observations.}

\item{theta}{A vector or matrix of parameters with \code{nparams} columns.}

\item{ncores}{If \code{model} is compiled with \code{OpenMP}, the number of cores to use for parallel processing.  Otherwise, uses \code{ncores = 1} and gives a warning.}
}
\value{
A vector of likelihood evaluations.  If input contains invalid data or parameters an error is thrown.
}
\description{
SDE Loglikelihood Function
}
