% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.make.model.R
\name{sde.make.model}
\alias{sde.make.model}
\title{Create an SDE model object.}
\usage{
sde.make.model(ModelFile, PriorFile = "default", data.names, param.names,
  hyper.check, OpenMP = FALSE, ...)
}
\arguments{
\item{ModelFile}{Path to the header file where the sde model is defined.}

\item{PriorFile}{Path to the header file where the prior is defined.  See \code{\link{sde.prior}} for details.}

\item{data.names}{Optional vector of names for the components of the sde.  Defaults to \code{X1,...,Xd}.}

\item{param.names}{Optional vector of names for the parameters of the sde.  Defaults to \code{theta1,...,thetap}.}

\item{hyper.check}{A function with arguments \code{prior.args}, \code{param.names}, and \code{data.names} used for passing the model hyper parameters to the C++ code.  See \code{\link{mvn.hyper.check}} for details.}

\item{OpenMP}{This sets whether the model is compiled for usage in \code{OpenMP}.}

\item{...}{additional parameters that are passed to \code{Rcpp::sourceCpp} when compiling the C++ code.}
}
\value{
An \code{sde.model} object, which is a list containing the following elements:
\describe{
\item{ndims, nparams}{The number of sde components and parameters.}
\item{data.names, param.names}{The names of the sde components and parameters.}
\item{drift, diff}{Functions to evaluate the drift and diffusion.}
\item{loglik}{The loglikelihood.}
\item{logprior}{The custom log prior.}
\item{sim}{Function for simulating data.}
\item{post}{MCMC sampler for posterior distribution.}
\item{omp}{a logical flag for whether or not the model was compiled for multicore functionality with \code{OpenMP}.}
}
}
\description{
Create an SDE model object.
}
\details{
The functions \code{sim}, \code{post}, \code{drift}, \code{diff}, \code{logpior}, and \code{loglik} should never be called directly. Instead use \code{sde.sim}, \code{sde.post} \code{sde.diff}, \code{sde.drift} and \code{sde.loglik}.

The code is compiled by copying the \code{ModelFile} to the \code{tmpdir} directory, along with a wrapper \code{.cpp} file to be compiled by \code{Rcpp::sourceCpp}.
}
\examples{
\donttest{
hex <- example.models("hest")
sde.make.model(ModelFile = hex$ModelFile,
               param.names = hex$param.names,
               data.names = hex$data.names)

}
}
