% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.pmcmc.R
\name{sde.pmcmc}
\alias{sde.pmcmc}
\title{Particle MCMC Sampling for SDE}
\usage{
sde.pmcmc(model, init, theta0, fixed.params, nsamples, npart, dT,
  resample = "multi", threshold = 0.5, mwg.sd = 1)
}
\arguments{
\item{model}{An \code{sde.model} object constructed with \code{\link{sde.make.model}}.}

\item{init}{An \code{sde.init} object constructed with \code{\link{sde.init}}.}

\item{theta0}{Initial parameter values.}

\item{fixed.params}{Logical vector of length \code{nparams} indicating which parameters are to be held fixed in the MCMC sampler.}

\item{nsamples}{Number of particle MCMC iterations.}

\item{npart}{Number of particles, fixed integer value.}

\item{dT}{Scalar interobservation time.}

\item{resample}{The type of particle resampling scheme to use. These are: multi(nomial), resid(ual), strat(ified), sys(tematic).}

\item{threshold}{A scalar less than 1 to indicate the threshold for resampling. A negative number disables resampling.}

\item{mwg.sd}{Standard deviation jump size for Metropolis-within-Gibbs on parameters and missing components of first SDE observation.}
}
\value{
A list of the following elements:
\describe{
  \item{\code{params}}{An \code{nsamples x nparams} matrix of posterior parameter draws.}
  \item{\code{accept}}{A named list of acceptance rates for the various components of the MCMC sampler.}
  \item{\code{init}}{The \code{sde.init} object which initialized the sampler.}
  \item{\code{mwg.sd}}{A named vector of Metropolis-within-Gibbs standard devations used at the last posterior iteration.}
}
}
\description{
Particle MCMC Sampling for SDE
}
