% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.post.R
\name{sde.post}
\alias{sde.post}
\title{Gibbs sampler for Multivariate SDEs}
\usage{
sde.post(model, init, hyper, nsamples, burn, mwg.sd = NULL, adapt = TRUE,
  loglik.out = FALSE, last.miss.out = FALSE, update.data = TRUE, data.out,
  update.params = TRUE, fixed.params, ncores = 1, verbose = TRUE,
  debug = FALSE)
}
\arguments{
\item{model}{An \code{sde.model} object constructed with \code{\link{sde.make.model}}.}

\item{init}{An \code{sde.init} object constructed with \code{\link{sde.init}}.}

\item{hyper}{The hyperparameters of the SDE prior.  See \code{\link{sde.prior}}.}

\item{nsamples}{Number of iterations.}

\item{burn}{Integer number of burnin samples, or fraction of \code{nsamples} to prepend as burnin.}

\item{mwg.sd}{standard deviation jump size for Metropolis-within-Gibbs steps (see Details).}

\item{adapt}{Logical or list to specify adaptive Metropolis-within-Gibbs sampling (see Details).}

\item{loglik.out}{Logical, whether to return the loglikelihood at each step.}

\item{last.miss.out}{Logical, whether to return the missing sde components of the last observation.}

\item{update.data}{Logical, whether to update the missing data.}

\item{data.out}{A scalar, integer vector, or list of three integer vectors determines the subset of data to be returned (see Details).}

\item{update.params}{Logical, whether to update the model parameters.}

\item{fixed.params}{Logical vector of length \code{nparams} indicating which parameters are to be held fixed in the MCMC sampler.}

\item{ncores}{If \code{model} is compiled with \code{OpenMP}, the number of cores to use for parallel processing.  Otherwise, uses \code{ncores = 1} and gives a warning.}

\item{verbose}{Logical, whether to periodically output MCMC status.}

\item{details}{The MWG jump sizes can be specified as a scalar, a vector or length \code{nparams + ndims}, or a named vector containing the elements defined by \code{nmiss0} and \code{fixed.params}.  The default jump sizes for each MWG random variable are \code{.25 * |initial_value|}.

\code{adapt = TRUE} implements an adaptive MCMC by Rosenthal and Roberts (2005).  At step \eqn{n} of the MCMC, the jump size of each MWG random variable is increased or decreased by \eqn{\delta(n)}, depending on whether the cumulative acceptance rate is above or below the optimal value of 0.44.  If \eqn{\sigma_n} is the size of the jump at step \eqn{n}, then the next jump size is determined by
\deqn{
\log(\sigma_{n+1}) = \log(\sigma_n) \pm \delta(n), \qquad \delta(n) = \min(.01, 1/n^{1/2}).
}
When \code{adapt} is not logical, it is a list with elements \code{max} and \code{rate}, such that \code{delta(n) = min(max, 1/n^rate)}.  These elements can be scalars or vectors in the same manner as \code{mwg.sd}.

For SDE models with thousands of latent variables, \code{data.out} can be used to thin the MCMC missing data output.  An integer vector or scalar returns specific or evenly-spaced posterior samples from the \code{ncomp x ndims} complete data matrix.  A list with elements \code{row} and \code{col} determines which samples to return with the former, and which of the \code{ncomp} time-points to return with the latter.}
}
\value{
A list with elements:
\describe{
  \item{params}{An \code{nsamples x nparams} matrix of posterior parameter draws.}
  \item{data}{A 3-d array of posterior missing data draws, for which the output dimensions are specified by \code{data.out}.}
  \item{data.out}{A list of three integer vectors specifying which timepoints, variables, and MCMC iterations correspond to the values in the \code{data} output.}
  \item{init}{The \code{sde.init} object which initialized the sampler.}
  \item{loglik}{If \code{loglik.out == TRUE}, the vector of \code{nsamples} complete data loglikelihoods calculated at each posterior sample.}
  \item{last.iter}{A list with elements \code{data} and \code{params} giving the last MCMC sample.  Useful for resuming the MCMC from that point.}
  \item{last.miss}{If \code{last.miss.out == TRUE}, a \code{nsamples x nmissN} matrix of all posterior draws for the missing data in the final observation.  Useful for SDE forecasting at future timepoints.}
  \item{accept}{A named list of acceptance rates for the various components of the MCMC sampler.}
}
}
\description{
Gibbs sampler for Multivariate SDEs
}
