% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.prior.R
\name{sde.prior}
\alias{sde.prior}
\title{SDE Prior Function}
\usage{
sde.prior(model, theta, x, hyper)
}
\arguments{
\item{model}{An \code{sde.model} object.}

\item{theta}{A vector or matrix of parameters with \code{nparams} columns.}

\item{x}{A vector or matrix of data with \code{ndims} columns.}

\item{hyper}{The hyperparameters of the SDE prior.  See Details.}
}
\value{
A vector of log-prior densities evaluated at the inputs.
}
\description{
SDE Prior Function
}
\details{
The prior is constructed at the \code{C++} level by defining a function (i.e., public member) \code{double logPrior(double *theta, double *x)} within the \code{sdePrior} class (see Examples).  At the \code{R} level, the \code{prior.spec} argument of \code{sde.make.model} is a function with arguments \code{hyper}, \code{param.names}, \code{data.names} used to convert \code{hyper} into a list of \code{NULL} or double-vectors which get passed on to the \code{C++} code.  This function can also be used to throw \code{R}-level errors to protect the \code{C++} code from invalid inputs, as is done for the default prior in \code{\link{mvn.prior.spec}}.
}
