% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.is.valid.R
\name{sde.valid.data}
\alias{sde.valid.data}
\title{Data validator}
\usage{
sde.valid.data(model, x, theta)
}
\arguments{
\item{model}{An \code{sde.model} object.}

\item{x}{A matrix of data.}

\item{theta}{A length \code{nparams} vector of parameter values.}
}
\description{
Data validator
}
