% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.is.valid.R
\name{sde.valid.params}
\alias{sde.valid.params}
\title{Parameter validator}
\usage{
sde.valid.params(model, theta)
}
\arguments{
\item{model}{An \code{sde.model} object.}

\item{theta}{A length \code{nparams} vector of parameter values.}
}
\description{
Parameter validator
}
