% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msde-package.R
\name{dgcir}
\alias{dgcir}
\title{gcir stationary distribution (regular scale)}
\arguments{
\item{x}{}

\item{gamma}{}

\item{mu}{}

\item{sigma}{}

\item{lambda}{}

\item{log}{}

\item{debug}{a boolean (\code{FALSE} by default) if set to \code{TRUE}, will cause the function to open a browser mid-call}
}
\value{
a numeric vector
}
\description{
<description>
}

