% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msde-package.R
\name{dhest}
\alias{dhest}
\title{Stationary Distribution of Heston Model Log-Return Differences}
\arguments{
\item{x}{}

\item{alpha}{}

\item{gamma}{}

\item{mu}{}

\item{beta}{}

\item{sigma}{}

\item{rho}{}

\item{dT}{}

\item{inter}{}

\item{n}{}

\item{LConst}{}

\item{debug}{a boolean (\code{FALSE} by default) if set to \code{TRUE}, will cause the function to open a browser mid-call}
}
\value{
numeric
}
\description{
these will roughly be centered at alpha*dT \cr
            requires Fourier inversion of the characteristic function, ie. numerical \cr
            complex integration.  currently done by grid method. \cr
            LConst allows fewer exp evaluations if alpha is held constant. \cr
}

