% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdeLists.R
\docType{data}
\name{hestList}
\alias{hestList}
\title{Heston Model Parameters}
\format{A list containing ndims, nparams, param.names, data.names, sdeDr, sdeDf, isValidData, isValidParams}
\usage{
hestList
}
\description{
List of paramaters that can be passed to sde.make.model to produce a hest model.
}
\keyword{datasets}

