% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdeHeaders.R
\docType{package}
\name{msdeHeaders}
\alias{msdeHeaders}
\alias{msdeHeaders-package}
\title{Simulation and Inference for Multivariate Stochastic Differential Equations.}
\description{
Simulation and Inference for Multivariate Stochastic Differential Equations.
}
