% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msde-package.R
\name{sde.diff}
\alias{sde.diff}
\title{Diffusion Function}
\arguments{
\item{model}{An object of the sde.model class generated using \code{sde.make.model}}

\item{x}{A matrix of data}

\item{theta}{A matrix of parameters}
}
\value{
A matrix containing the result of the diffusion function
}
\description{
Applies the diffusion function for \code{model} to \code{x} using \code{theta}
}
\examples{
# Create the model
hest.model <- sde.make.model(list = hestList, model.name = "hest")

theta <- c(alpha = .1, gamma = 5, beta = .8, sigma = .6, rho = -.7)
Y0 <- c(X = log(100), Z = .1)

# simulate data
N <- 10
burn <- 10
dT <- 1/252

hsim <- sde.sim(model = hest.model, init.data = Y0, params = theta, dt = dT, dt.sim = dT/100,
               N = N, burn = burn, nreps = 1)

diff <- sde.diff(model = hest.model, x = hsim$data, theta = theta)
}

