% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msde-package.R
\name{sde.loglik}
\alias{sde.loglik}
\title{Loglikelihood Function}
\arguments{
\item{model}{An object of the sde.model class generated using \code{sde.make.model}}

\item{x}{A matrix of size (nreps x ncomp x ndims) containing data}

\item{dt}{A numeric vector of size 1 or ncomp-1 containing Delta-t}

\item{theta}{A matrix of size (nreps x nparams) containing parameters}

\item{debug}{A boolean (\code{FALSE} by default) if set to \code{TRUE}, will cause the function to open a browser mid-call}
}
\value{
a matrix containing the result of the log-likelihood function
}
\description{
Applies the log-likelihood function for \code{model} to \code{x} using \code{theta} and \code{dt}
}
\examples{
# Create the model
hest.model <- sde.make.model(list = hestList, model.name = "hest")

theta <- c(alpha = .1, gamma = 5, beta = .8, sigma = .6, rho = -.7)
Y0 <- c(X = log(100), Z = .1)

# simulate data
N <- 10
burn <- 10
dT <- 1/252

hsim <- sde.sim(model = hest.model, init.data = Y0, params = theta, dt = dT, dt.sim = dT/100,
               N = N, burn = burn, nreps = 1)

loglik <- sde.loglik(model = hest.model, x = hsim$data, dt = dT, theta = theta)
}

