% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.make.model.R
\name{sde.make.model}
\alias{sde.make.model}
\title{Create an SDE model object.}
\usage{
sde.make.model(ModelFile, PriorFile = "default", data.names, param.names,
  hyper.check, OpenMP = FALSE, ..., debug = FALSE)
}
\arguments{
\item{ModelFile}{Path to the header file where the SDE model is defined.}

\item{PriorFile}{Path to the header file where the SDE prior is defined.  See \code{\link{sde.prior}} for details.}

\item{data.names}{Vector of names for the SDE components.  Defaults to \code{X1,...,Xd}.}

\item{param.names}{Vector of names for the SDE parameters.  Defaults to \code{theta1,...,thetap}.}

\item{hyper.check}{A function with arguments \code{hyper}, \code{param.names}, and \code{data.names} used for passing the model hyper parameters to the C++ code.  See \code{\link{mvn.hyper.check}} for details.}

\item{OpenMP}{Logical; whether the model is compiled with \code{OpenMP} for C++ level parallelization.}

\item{...}{additional arguments to \code{Rcpp::sourceCpp} for compiling the C++ code.}
}
\value{
An \code{sde.model} object, consisting of a list with the following elements:
\describe{
\item{\code{model.ptr}}{Pointer to C++ sde object (\code{sdeCobj}) implementing the member functions: drift/diffusion, data/parameter validators, loglikelihood, prior distribution, forward simulation, MCMC algorithm for Bayesian inference.}
\item{\code{ndims, nparams}}{The number of SDE components and parameters.}
\item{\code{data.names, param.names}}{The names of the SDE components and parameters.}
\item{\code{omp}}{A logical flag for whether or not the model was compiled for multicore functionality with \code{OpenMP}.}
}
}
\description{
Compiles the C++ code for various SDE-related algorithms and makes the routines available within R.
}
\seealso{
\link{\code{sde.drift}}, \link{\code{sde.diff}}, \link{\code{sde.is.valid}}, \link{\code{sde.loglik}}, \link{\code{sde.prior}}, \link{\code{sde.sim}}, \link{\code{sde.post}}
}
