% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.pf.R
\name{sde.pf}
\alias{sde.pf}
\title{Particle filter for SDE model
which is currently restricted for eou model and level-1 Euler approximation, later we will make it to be more generalized.}
\usage{
sde.pf(model, theta, Y0, nObs, dt, dt.sim, nPart, verbose = FALSE)
}
\arguments{
\item{model}{An \code{sde.model} object.}

\item{theta}{A vector or matrix of size \code{nreps x nparams} of SDE parameters.}

\item{Y0}{A vector or a matrix of size \code{nreps x ndims} of the SDE values at time 0.}

\item{nObs}{The number of SDE observations per trajectory to generate.}

\item{dt}{Scalar interobservation time.}

\item{dt.sim}{Scalar interobservation time for simulation.  That is, interally the interobservation time is \code{dt.sim} but only one out of every \code{dt/dt.sim} simulation steps is kept in the output.}

\item{nPart}{Number of particles.}

\item{verbose}{Whether or not to display information on the simulation.}
}
\value{
A list with elements:
\describe{
  \item{\code{Yup}}{A matrix of all the historical data}
  \item{\code{lgwt}}{A matrix of all the historical normalized log weights.}
}
}
\description{
Simulate a particle filter based on discretized Euler-Maruyama approximation of a given SDE model
}
\details{
...
}
\examples{
# load pre-compiled model
emod <- sde.examples("eou")

# initialize the sde model
theta <- c(alpha = .1, gamma = 4.8, eta = 0.1, sigma = .1, rho = -.63) # true parameter values
nObs <- 100 # number of observations
dt <- 1/252 # time between observations (1 year has about 252 trading days)
dt.sim <- dt/10
Y0 <- c(X = rnorm(1), V = rnorm(1)) # initial SDE values
nPart <- 50

output <- sde.pf(emod, theta, Y0, nObs, dt, dt.sim, nPart, verbose = FALSE)

}
