% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.is.valid.R
\name{sde.valid.data}
\alias{sde.valid.data}
\title{Data validator}
\usage{
sde.valid.data(model, x, theta)
}
\arguments{
\item{model}{An \code{sde.model} object.}

\item{x}{A matrix of data.}

\item{theta}{A length \code{nparams} vector of parameter values.}
}
\description{
Data validator
}
\examples{
\donttest{
hex <- example.models("hest")
hmod <- sde.make.model(ModelFile = hex$ModelFile,
                       param.names = hex$param.names,
                       data.names = hex$data.names)

theta <- c(alpha = 0.1, gamma = 1, beta = 0.8, sigma = 0.6, rho = -0.8)

# example of not valid data
x0 <- c(X = log(1000), Z = -0.1)
sde.valid.data(model = hmod, x = x0, theta = theta)

# example of valid data
x0 <- c(X = log(1000), Z = 0.1)
sde.valid.data(model = hmod, x = x0, theta = theta)
}
}
