% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.is.valid.R
\name{sde.valid.params}
\alias{sde.valid.params}
\title{Parameter validator}
\usage{
sde.valid.params(model, theta)
}
\arguments{
\item{model}{An \code{sde.model} object.}

\item{theta}{A length \code{nparams} vector of parameter values.}
}
\description{
Parameter validator
}
\examples{
\donttest{
hex <- example.models("hest")
hmod <- sde.make.model(ModelFile = hex$ModelFile,
                       param.names = hex$param.names,
                       data.names = hex$data.names)

# example of not valid param
theta <- c(alpha = 0.1, gamma = -4, beta = 0.8, sigma = 0.6, rho = -0.8)
sde.valid.params(model = hmod, theta = theta)

# example of valid param
theta <- c(alpha = 0.1, gamma = 1, beta = 0.8, sigma = 0.6, rho = -0.8)
sde.valid.params(model = hmod, theta = theta)
}
}
