% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.init.R
\name{sv.init}
\alias{sv.init}
\title{Initialize Missing Volatility}
\usage{
sv.init(x, dt, vblock.size, debug = FALSE)
}
\arguments{
\item{debug}{a boolean (\code{FALSE} by default) if set to \code{TRUE}, will cause the function to open a browser mid-call}
}
\value{
a list containing: alpha, v
}

